#551 reorder tests
